+++ src/Data/MediaBus/Rtp/Source.hs
@@ -1,12 +1,13 @@
--- | A conduit 'Source' that receives bytestrings and parses RTP packets.
 {-# OPTIONS -Wno-unused-top-binds #-}
 
+-- | A conduit 'Source' that receives bytestrings and parses RTP packets.
 module Data.MediaBus.Rtp.Source
-  ( type RtpStream
-  , rtpSource
-  , rtpPayloadDemux
-  , type RtpPayloadHandler
-  ) where
+  ( type RtpStream,
+    rtpSource,
+    rtpPayloadDemux,
+    type RtpPayloadHandler,
+  )
+where
 
 import Conduit
 import Control.Lens
@@ -17,8 +18,8 @@ import Data.Default
 import qualified Data.List
 import Data.Maybe
 import Data.MediaBus
+import Data.MediaBus.Rtp.InternalLogging
 import Data.MediaBus.Rtp.Packet
-import Data.String
 import Text.Printf
 
 -- | A handy alias for a 'Stream' for RTP packets, with according sequence
@@ -26,30 +27,31 @@ import Text.Printf
 type RtpStream p = Stream RtpSsrc RtpSeqNum RtpTimestamp p RtpPayload
 
 data RRState ctx = MkRRState
-  { _currCtx :: ctx
-  , _isFirstPacket :: Bool
-  } deriving (Show)
+  { _currCtx :: ctx,
+    _isFirstPacket :: Bool
+  }
+  deriving (Show)
 
 makeLenses ''RRState
 
 -- | A 'Conduit' that consumes 'ByteString's and produces an 'RtpStream'
-rtpSource
-  :: ( Show i
-     , Show s
-     , Show t
-     , Show p
-     , Default i
-     , Monad m
-     , Default p
-     , MonadLogger m
-     )
-  => ConduitT (Stream i s t p B.ByteString) (RtpStream p) m ()
+rtpSource ::
+  ( Show i,
+    Show s,
+    Show t,
+    Show p,
+    Default i,
+    Monad m,
+    Default p,
+    MonadLogger m
+  ) =>
+  ConduitT (Stream i s t p B.ByteString) (RtpStream p) m ()
 rtpSource =
   evalStateC (MkRRState (MkFrameCtx def def def def) True) $
-  awaitForever processFrames
+    awaitForever processFrames
   where
     processFrames frm@(MkStream (Start _)) =
-      $logInfo (fromString ("start frame received: " ++ show frm))
+      out ("start frame received: " ++ show frm)
     processFrames (MkStream (Next (MkFrame _ _ !contentIn))) =
       case deserialize contentIn of
         Left rtpError -> logRtpError rtpError
@@ -60,28 +62,32 @@ rtpSource =
           yieldStreamNext (body rtpPacket)
     logRtpError e = do
       ctx <- use currCtx
-      $logError (fromString (printf "rtp packet parse error: %s, frame-context: %s" e (show ctx)))
+      err (printf "rtp packet parse error: %s, frame-context: %s" e (show ctx))
     updateState rtpHeader = do
       oldCtx <-
-        currCtx <<%=
-        ((frameCtxSeqNumRef .~ sequenceNumber rtpHeader) .
-         (frameCtxTimestampRef .~ headerTimestamp rtpHeader))
+        currCtx
+          <<%= ( (frameCtxSeqNumRef .~ sequenceNumber rtpHeader)
+                   . (frameCtxTimestampRef .~ headerTimestamp rtpHeader)
+               )
       wasFirstPacket <- isFirstPacket <<.= False
       if oldCtx ^. frameCtxSourceId /= ssrc rtpHeader
         then do
           currCtx . frameCtxSourceId .= ssrc rtpHeader
           return RtpSsrcChanged
-        else if sequenceNumbersDifferTooMuch
-                  (oldCtx ^. frameCtxSeqNumRef)
-                  (sequenceNumber rtpHeader)
-               then return RtpSequenceNumberGap
-               else if timestampsDifferTooMuch
-                         (oldCtx ^. frameCtxTimestampRef)
-                         (headerTimestamp rtpHeader)
-                      then return RtpTimestampGap
-                      else if wasFirstPacket
-                             then return NewRtpSession
-                             else return FrameCtxNotChanged
+        else
+          if sequenceNumbersDifferTooMuch
+            (oldCtx ^. frameCtxSeqNumRef)
+            (sequenceNumber rtpHeader)
+            then return RtpSequenceNumberGap
+            else
+              if timestampsDifferTooMuch
+                (oldCtx ^. frameCtxTimestampRef)
+                (headerTimestamp rtpHeader)
+                then return RtpTimestampGap
+                else
+                  if wasFirstPacket
+                    then return NewRtpSession
+                    else return FrameCtxNotChanged
       where
         sequenceNumbersDifferTooMuch oldSN currSN =
           let d =
@@ -89,17 +95,17 @@ rtpSource =
                   then currSN - oldSN
                   else oldSN - currSN -- TODO use LocalOrd??
               sequenceNumberMaxDelta = 10
-          in d >= sequenceNumberMaxDelta
+           in d >= sequenceNumberMaxDelta
         timestampsDifferTooMuch oldTS currTS =
           let d =
                 if currTS >= oldTS
                   then currTS - oldTS
                   else oldTS - currTS
               timestampMaxDelta = 2000 -- TODO extract
-          in d >= timestampMaxDelta
+           in d >= timestampMaxDelta
     yieldStreamStart why = do
       fx <- use currCtx
-      $logDebug (fromString (show why ++ ": starting rtp stream: " ++ show fx))
+      dbg (show why ++ ": starting rtp stream: " ++ show fx)
       yieldStartFrameCtx fx
     yieldStreamNext p = do
       ts <- use (currCtx . frameCtxTimestampRef)
@@ -117,14 +123,14 @@ data RRSourceChange
 
 -- | A utility that call the right 'RtpPayloadHandler' for the 'RtpPayloadType'
 -- of the 'Frame'.
-rtpPayloadDemux
-  :: (Integral t, Monad m)
-  => [(RtpPayloadType, RtpPayloadHandler (Ticks r t) c)]
-  -> c
-  -> ConduitT (RtpStream p) (Stream RtpSsrc RtpSeqNum (Ticks r t) p c) m ()
+rtpPayloadDemux ::
+  (Integral t, Monad m) =>
+  [(RtpPayloadType, RtpPayloadHandler (Ticks r t) c)] ->
+  c ->
+  ConduitT (RtpStream p) (Stream RtpSsrc RtpSeqNum (Ticks r t) p c) m ()
 rtpPayloadDemux payloadTable fallbackContent =
-  mapC (timestamp %~ (MkTicks . fromIntegral . _rtpTimestamp)) .|
-  awaitForever go
+  mapC (timestamp %~ (MkTicks . fromIntegral . _rtpTimestamp))
+    .| awaitForever go
   where
     setFallbackContent = framePayload .~ fallbackContent
     go (MkStream (Next !frm)) = do